[tweak] don’t derp out if there is no application for some reason
SpringBoard can pass nil as the application to open, and it’ll try and guess
the app for you. Messages notifications from the notification center and the web
search button in Spotlight both do this, so the open url request pretty much got
cancelled. Now if there’s no application provided, it falls back to the one that
was in the original open url request (which might not even have provided an app,
but that’s ok because SpringBoard’s got us covered).